Further refactor, although it now appears we just care about the .csv for the usage corpus
